+++ src/main/java/dev/lydtech/component/framework/extension/DockerManager.java
@@ -103,7 +103,6 @@ public final class DockerManager {
         findContainerAndMapPort(dockerClient, KAFKA.toString(), KAFKA_ENABLED, KAFKA_PORT);
         findContainerAndMapPort(dockerClient, DEBEZIUM.toString(), DEBEZIUM_ENABLED, DEBEZIUM_PORT);
         findContainerAndMapPort(dockerClient, KAFKA_SCHEMA_REGISTRY.toString().replace("_", "."), KAFKA_SCHEMA_REGISTRY_ENABLED, KAFKA_SCHEMA_REGISTRY_PORT);
-        findContainerAndMapPort(dockerClient, KAFKA_CONTROL_CENTER.toString().replace("_", "."), KAFKA_CONTROL_CENTER_ENABLED, KAFKA_CONTROL_CENTER_PORT);
         findContainerAndMapPort(dockerClient, WIREMOCK.toString(), WIREMOCK_ENABLED, WIREMOCK_PORT);
         findContainerAndMapPort(dockerClient, LOCALSTACK.toString(), LOCALSTACK_ENABLED, LOCALSTACK_PORT);